refactor: 출석기록 이름 검사 메서드 클래스 이동
